# Conflicts:
#	controllers/servers/csi/addons_server.py
#	controllers/tests/array_action/svc/array_mediator_svc_test.py
#	controllers/tests/controller_server/addons_server_test.py

+++ controllers/servers/host_definer/kubernetes_manager/manager.py
@@ -1,4 +1,5 @@
 import datetime
+import base64
 
 from kubernetes import client, config, dynamic
 from kubernetes.client import api_client
@@ -7,6 +8,7 @@ from kubernetes.client.rest import ApiException
 from controllers.common.csi_logger import get_stdout_logger
 import controllers.servers.host_definer.messages as messages
 from controllers.servers.host_definer import settings
+import controllers.common.settings as common_settings
 from controllers.servers.host_definer.types import (
     CsiNodeInfo, PodInfo, NodeInfo, StorageClassInfo, HostDefinitionInfo)
 
@@ -44,6 +46,7 @@ class KubernetesManager():
         for k8s_csi_node in k8s_csi_nodes:
             if self._is_k8s_csi_node_has_driver(k8s_csi_node):
                 csi_nodes_info_with_driver.append(self._generate_csi_node_info(k8s_csi_node))
+        logger.info(messages.CSI_NODES_WITH_IBM_BLOCK_CSI_DRIVER.format(csi_nodes_info_with_driver))
         return csi_nodes_info_with_driver
 
     def _get_k8s_csi_nodes(self):
@@ -63,11 +66,6 @@ class KubernetesManager():
         except ApiException as ex:
             logger.error(messages.FAILED_TO_GET_NODES.format(ex.body))
             return []
-
-    def _generate_node_info(self, k8s_node):
-        node_info = NodeInfo()
-        node_info.name = k8s_node.metadata.name
-        return node_info
 
     def _get_storage_classes_info(self):
         try:
@@ -91,7 +89,8 @@ class KubernetesManager():
     def _is_k8s_csi_node_has_driver(self, k8s_csi_node):
         if k8s_csi_node.spec.drivers:
             for driver in k8s_csi_node.spec.drivers:
-                return driver.name == settings.CSI_PROVISIONER_NAME
+                if driver.name == settings.CSI_PROVISIONER_NAME:
+                    return True
         return False
 
     def _get_csi_node_info(self, node_name):
@@ -99,7 +98,7 @@ class KubernetesManager():
             k8s_csi_node = self.csi_nodes_api.get(name=node_name)
             return self._generate_csi_node_info(k8s_csi_node)
         except ApiException as ex:
-            if ex.status != 404:
+            if ex.status == 404:
                 logger.error(messages.CSI_NODE_DOES_NOT_EXIST.format(node_name))
             else:
                 logger.error(messages.FAILED_TO_GET_CSI_NODE.format(node_name, ex.body))
@@ -116,7 +115,7 @@ class KubernetesManager():
             for driver in k8s_csi_node.spec.drivers:
                 if driver.name == settings.CSI_PROVISIONER_NAME:
                     return driver.nodeID
-        return None
+        return ''
 
     def _get_matching_host_definition_info(self, node_name, secret_name, secret_namespace):
         k8s_host_definitions = self._get_k8s_host_definitions()
@@ -137,7 +136,7 @@ class KubernetesManager():
     def _generate_host_definition_info(self, k8s_host_definition):
         host_definition_info = HostDefinitionInfo()
         host_definition_info.name = k8s_host_definition.metadata.name
-        host_definition_info.resource_version = self._get_host_definition_resource_version(k8s_host_definition)
+        host_definition_info.resource_version = self._get_k8s_object_resource_version(k8s_host_definition)
         host_definition_info.uid = k8s_host_definition.metadata.uid
         host_definition_info.phase = self._get_host_definition_phase(k8s_host_definition)
         host_definition_info.secret_name = self._get_attr_from_host_definition(
@@ -146,13 +145,16 @@ class KubernetesManager():
             k8s_host_definition, settings.SECRET_NAMESPACE_FIELD)
         host_definition_info.node_name = self._get_attr_from_host_definition(
             k8s_host_definition, settings.NODE_NAME_FIELD)
-        host_definition_info.node_id = self._get_attr_from_host_definition(k8s_host_definition, settings.NODE_ID_FIELD)
+        host_definition_info.node_id = self._get_attr_from_host_definition(
+            k8s_host_definition, common_settings.HOST_DEFINITION_NODE_ID_FIELD)
+        host_definition_info.connectivity_type = self._get_attr_from_host_definition(
+            k8s_host_definition, settings.CONNECTIVITY_TYPE_FIELD)
         return host_definition_info
 
-    def _get_host_definition_resource_version(self, k8s_host_definition):
-        if k8s_host_definition.metadata.resource_version:
-            return k8s_host_definition.metadata.resource_version
-        return k8s_host_definition.metadata.resourceVersion
+    def _get_k8s_object_resource_version(self, k8s_object):
+        if k8s_object.metadata.resource_version:
+            return k8s_object.metadata.resource_version
+        return k8s_object.metadata.resourceVersion
 
     def _get_host_definition_phase(self, k8s_host_definition):
         if k8s_host_definition.status:
@@ -172,24 +174,25 @@ class KubernetesManager():
     def _create_host_definition(self, host_definition_manifest):
         try:
             k8s_host_definition = self.host_definitions_api.create(body=host_definition_manifest)
+            logger.info(messages.CREATED_HOST_DEFINITION.format(k8s_host_definition.metadata.name))
             self._add_finalizer(k8s_host_definition.metadata.name)
             return self._generate_host_definition_info(k8s_host_definition)
         except ApiException as ex:
             if ex != 404:
                 logger.error(messages.FAILED_TO_CREATE_HOST_DEFINITION.format(
-                    host_definition_manifest[settings.METADATA][settings.NAME], ex.body))
+                    host_definition_manifest[settings.METADATA][common_settings.NAME_FIELD], ex.body))
 
     def _add_finalizer(self, host_definition_name):
         logger.info(messages.ADD_FINALIZER_TO_HOST_DEFINITION.format(host_definition_name))
-        self._update_finalizer(
-            host_definition_name, [settings.HOST_DEFINITION_PLURAL + '.' + settings.CSI_IBM_GROUP, ])
+        self._update_finalizer(host_definition_name, [settings.CSI_IBM_FINALIZER, ])
 
     def _set_host_definition_status(self, host_definition_name, host_definition_phase):
         logger.info(messages.SET_HOST_DEFINITION_STATUS.format(host_definition_name, host_definition_phase))
         status = self._get_status_manifest(host_definition_phase)
         try:
             self.custom_object_api.patch_cluster_custom_object_status(
-                settings.CSI_IBM_GROUP, settings.VERSION, settings.HOST_DEFINITION_PLURAL, host_definition_name, status)
+                common_settings.CSI_IBM_GROUP, common_settings.VERSION, common_settings.HOST_DEFINITION_PLURAL,
+                host_definition_name, status)
         except ApiException as ex:
             if ex.status == 404:
                 logger.error(messages.HOST_DEFINITION_DOES_NOT_EXIST.format(host_definition_name))
@@ -197,13 +200,11 @@ class KubernetesManager():
                 logger.error(messages.FAILED_TO_SET_HOST_DEFINITION_STATUS.format(host_definition_name, ex.body))
 
     def _get_status_manifest(self, host_definition_phase):
-        status = {
+        return {
             settings.STATUS: {
                 settings.PHASE: host_definition_phase,
             }
         }
-
-        return status
 
     def _generate_k8s_event(self, host_definition_info, message, action, message_type):
         return client.CoreV1Event(
@@ -231,36 +232,42 @@ class KubernetesManager():
     def _delete_host_definition(self, host_definition_name):
         logger.info(messages.DELETE_HOST_DEFINITION.format(host_definition_name))
         try:
-            self._remove_finalizer(host_definition_name)
-            self.host_definitions_api.delete(name=host_definition_name, body={})
+            remove_finalizer_status_code = self._remove_finalizer(host_definition_name)
+            if remove_finalizer_status_code == 200:
+                self.host_definitions_api.delete(name=host_definition_name, body={})
+            else:
+                logger.error(messages.FAILED_TO_DELETE_HOST_DEFINITION.format(
+                    host_definition_name, messages.FAILED_TO_REMOVE_FINALIZER))
         except ApiException as ex:
             if ex.status != 404:
                 logger.error(messages.FAILED_TO_DELETE_HOST_DEFINITION.format(host_definition_name, ex.body))
 
     def _remove_finalizer(self, host_definition_name):
         logger.info(messages.REMOVE_FINALIZER_TO_HOST_DEFINITION.format(host_definition_name))
-        self._update_finalizer(host_definition_name, [])
+        return self._update_finalizer(host_definition_name, [])
 
     def _update_finalizer(self, host_definition_name, finalizers):
         finalizer_manifest = {
             settings.METADATA: {
-                settings.NAME: host_definition_name,
+                common_settings.NAME_FIELD: host_definition_name,
                 settings.FINALIZERS: finalizers,
             }
         }
-        self._patch_host_definition(finalizer_manifest)
+        return self._patch_host_definition(finalizer_manifest)
 
     def _patch_host_definition(self, host_definition_manifest):
-        host_definition_name = host_definition_manifest[settings.METADATA][settings.NAME]
+        host_definition_name = host_definition_manifest[settings.METADATA][common_settings.NAME_FIELD]
         logger.info(messages.PATCHING_HOST_DEFINITION.format(host_definition_name))
         try:
             self.host_definitions_api.patch(body=host_definition_manifest, name=host_definition_name,
                                             content_type='application/merge-patch+json')
+            return 200
         except ApiException as ex:
             if ex.status == 404:
                 logger.error(messages.HOST_DEFINITION_DOES_NOT_EXIST.format(host_definition_name))
             else:
                 logger.error(messages.FAILED_TO_PATCH_HOST_DEFINITION.format(host_definition_name, ex.body))
+            return ex.status
 
     def _update_manage_node_label(self, node_name, label_value):
         body = self._get_body_for_labels(label_value)
@@ -280,22 +287,55 @@ class KubernetesManager():
 
         return body
 
-    def _get_data_from_secret(self, secret_name, secret_namespace):
+    def _get_secret_data(self, secret_name, secret_namespace):
         try:
-            return self.core_api.read_namespaced_secret(name=secret_name, namespace=secret_namespace).data
+            logger.info(messages.READ_SECRET.format(secret_name, secret_namespace))
+            secret_data = self.core_api.read_namespaced_secret(name=secret_name, namespace=secret_namespace).data
+            return self._change_decode_base64_secret_config(secret_data)
         except ApiException as ex:
             if ex.status == 404:
                 logger.error(messages.SECRET_DOES_NOT_EXIST.format(secret_name, secret_namespace))
             else:
                 logger.error(messages.FAILED_TO_GET_SECRET.format(secret_name, secret_namespace, ex.body))
-            return None
+        return {}
+
+    def _change_decode_base64_secret_config(self, secret_data):
+        if settings.SECRET_CONFIG_FIELD in secret_data.keys():
+            secret_data[settings.SECRET_CONFIG_FIELD] = self._decode_base64_to_dict(
+                secret_data[settings.SECRET_CONFIG_FIELD])
+        return secret_data
+
+    def _decode_base64_to_dict(self, content_with_base64):
+        decoded_string_content = self._decode_base64_to_string(content_with_base64)
+        encoded_dict = str(decoded_string_content).encode('utf-8')
+        base64_dict = base64.b64encode(encoded_dict)
+        my_dict_again = eval(base64.b64decode(base64_dict))
+        return my_dict_again
+
+    def _decode_base64_to_string(self, content_with_base64):
+        try:
+            base64_bytes = content_with_base64.encode('ascii')
+            decoded_string_in_bytes = base64.b64decode(base64_bytes)
+            return decoded_string_in_bytes.decode('ascii')
+        except Exception:
+            return content_with_base64
+
+    def _get_node_info(self, node_name):
+        k8s_node = self._read_node(node_name)
+        if k8s_node:
+            return self._generate_node_info(k8s_node)
+        return NodeInfo('', {})
 
     def _read_node(self, node_name):
         try:
+            logger.info(messages.READ_NODE.format(node_name))
             return self.core_api.read_node(name=node_name)
         except ApiException as ex:
             logger.error(messages.FAILED_TO_GET_NODE.format(node_name, ex.body))
             return None
+
+    def _generate_node_info(self, k8s_node):
+        return NodeInfo(k8s_node.metadata.name, k8s_node.metadata.labels)
 
     def _get_csi_daemon_set(self):
         try:
@@ -318,7 +358,7 @@ class KubernetesManager():
 
         except ApiException as ex:
             logger.error(messages.FAILED_TO_LIST_PODS.format(ex.body))
-            return None
+            return []
 
     def _generate_pod_info(self, k8s_pod):
         pod_info = PodInfo()
